Refactor package structure for RabbitMQ and Redis connections
