+++ src/main/java/fr/milekat/utils/messaging/adapter/redis/RedisConnection.java
@@ -1,0 +1,4 @@
+package fr.milekat.utils.messaging.adapter.redis;
+
+public class RedisConnection {
+}
