--- src/main/java/fr/milekat/utils/messaging/adapter/redis/connection/RedisConnection.java
@@ -1,4 +1,0 @@
-package fr.milekat.utils.messaging.adapter.redis.connection;
-
-public class RedisConnection {
-}
